+++ opencascade-hs/src/OpenCascade/STEPControl/Types.hs
@@ -1,6 +1,8 @@
 {-# LANGUAGE EmptyDataDecls #-}
 module OpenCascade.STEPControl.Types
 ( Writer
+, Reader
 ) where
 
 data Writer
+data Reader