+++ opencascade-hs/src/OpenCascade/XSControl/Types.hs
@@ -1,0 +1,6 @@
+{-# LANGUAGE EmptyDataDecls #-}
+module OpenCascade.XSControl.Types
+( Reader
+) where
+
+data Reader